use int builder
