+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/aggregation/X-ValuesAggregator.java.st
@@ -102,82 +102,7 @@ $if(BytesRef)$
         GroupingState state,
         BytesRefBlock values
     ) {
-        OrdinalBytesRefBlock valuesOrdinal = values.asOrdinals();
-        if (valuesOrdinal == null) {
-            return delegate;
-        }
-        BytesRefVector dict = valuesOrdinal.getDictionaryVector();
-        final int[] hashIds = new int[dict.getPositionCount()];
-        BytesRef spare = new BytesRef();
-        for (int p = 0; p < dict.getPositionCount(); p++) {
-            hashIds[p] = Math.toIntExact(BlockHash.hashOrdToGroup(state.bytes.add(dict.getBytesRef(p, spare))));
-        }
-        IntBlock ordinalIds = valuesOrdinal.getOrdinalsBlock();
-        return new GroupingAggregatorFunction.AddInput() {
-            @Override
-            public void add(int positionOffset, IntArrayBlock groupIds) {
-                for (int groupPosition = 0; groupPosition < groupIds.getPositionCount(); groupPosition++) {
-                    if (groupIds.isNull(groupPosition)) {
-                        continue;
-                    }
-                    int groupStart = groupIds.getFirstValueIndex(groupPosition);
-                    int groupEnd = groupStart + groupIds.getValueCount(groupPosition);
-                    for (int g = groupStart; g < groupEnd; g++) {
-                        int groupId = groupIds.getInt(g);
-                        if (ordinalIds.isNull(groupPosition + positionOffset)) {
-                            continue;
-                        }
-                        int valuesStart = ordinalIds.getFirstValueIndex(groupPosition + positionOffset);
-                        int valuesEnd = valuesStart + ordinalIds.getValueCount(groupPosition + positionOffset);
-                        for (int v = valuesStart; v < valuesEnd; v++) {
-                            state.values.add(groupId, ordinalIds.getInt(v));
-                        }
-                    }
-                }
-            }
-
-            @Override
-            public void add(int positionOffset, IntBigArrayBlock groupIds) {
-                for (int groupPosition = 0; groupPosition < groupIds.getPositionCount(); groupPosition++) {
-                    if (groupIds.isNull(groupPosition)) {
-                        continue;
-                    }
-                    int groupStart = groupIds.getFirstValueIndex(groupPosition);
-                    int groupEnd = groupStart + groupIds.getValueCount(groupPosition);
-                    for (int g = groupStart; g < groupEnd; g++) {
-                        int groupId = groupIds.getInt(g);
-                        if (ordinalIds.isNull(groupPosition + positionOffset)) {
-                            continue;
-                        }
-                        int valuesStart = ordinalIds.getFirstValueIndex(groupPosition + positionOffset);
-                        int valuesEnd = valuesStart + ordinalIds.getValueCount(groupPosition + positionOffset);
-                        for (int v = valuesStart; v < valuesEnd; v++) {
-                            state.values.add(groupId, ordinalIds.getInt(v));
-                        }
-                    }
-                }
-            }
-
-            @Override
-            public void add(int positionOffset, IntVector groupIds) {
-                for (int groupPosition = 0; groupPosition < groupIds.getPositionCount(); groupPosition++) {
-                    int groupId = groupIds.getInt(groupPosition);
-                    if (ordinalIds.isNull(groupPosition + positionOffset)) {
-                        continue;
-                    }
-                    int valuesStart = ordinalIds.getFirstValueIndex(groupPosition + positionOffset);
-                    int valuesEnd = valuesStart + ordinalIds.getValueCount(groupPosition + positionOffset);
-                    for (int v = valuesStart; v < valuesEnd; v++) {
-                        state.values.add(groupId, ordinalIds.getInt(v));
-                    }
-                }
-            }
-
-            @Override
-            public void close() {
-                delegate.close();
-            }
-        };
+        return ValuesBytesRefAggregators.wrapAddInput(delegate, state, values);
     }
 
     public static GroupingAggregatorFunction.AddInput wrapAddInput(
@@ -185,61 +110,7 @@ $if(BytesRef)$
         GroupingState state,
         BytesRefVector values
     ) {
-        var valuesOrdinal = values.asOrdinals();
-        if (valuesOrdinal == null) {
-            return delegate;
-        }
-        BytesRefVector dict = valuesOrdinal.getDictionaryVector();
-        final int[] hashIds = new int[dict.getPositionCount()];
-        BytesRef spare = new BytesRef();
-        for (int p = 0; p < dict.getPositionCount(); p++) {
-            hashIds[p] = Math.toIntExact(BlockHash.hashOrdToGroup(state.bytes.add(dict.getBytesRef(p, spare))));
-        }
-        var ordinalIds = valuesOrdinal.getOrdinalsVector();
-        return new GroupingAggregatorFunction.AddInput() {
-            @Override
-            public void add(int positionOffset, IntArrayBlock groupIds) {
-                for (int groupPosition = 0; groupPosition < groupIds.getPositionCount(); groupPosition++) {
-                    if (groupIds.isNull(groupPosition)) {
-                        continue;
-                    }
-                    int groupStart = groupIds.getFirstValueIndex(groupPosition);
-                    int groupEnd = groupStart + groupIds.getValueCount(groupPosition);
-                    for (int g = groupStart; g < groupEnd; g++) {
-                        int groupId = groupIds.getInt(g);
-                        state.values.add(groupId, ordinalIds.getInt(groupPosition + positionOffset));
-                    }
-                }
-            }
-
-            @Override
-            public void add(int positionOffset, IntBigArrayBlock groupIds) {
-                for (int groupPosition = 0; groupPosition < groupIds.getPositionCount(); groupPosition++) {
-                    if (groupIds.isNull(groupPosition)) {
-                        continue;
-                    }
-                    int groupStart = groupIds.getFirstValueIndex(groupPosition);
-                    int groupEnd = groupStart + groupIds.getValueCount(groupPosition);
-                    for (int g = groupStart; g < groupEnd; g++) {
-                        int groupId = groupIds.getInt(g);
-                        state.values.add(groupId, ordinalIds.getInt(groupPosition + positionOffset));
-                    }
-                }
-            }
-
-            @Override
-            public void add(int positionOffset, IntVector groupIds) {
-                for (int groupPosition = 0; groupPosition < groupIds.getPositionCount(); groupPosition++) {
-                    int groupId = groupIds.getInt(groupPosition);
-                    state.values.add(groupId, ordinalIds.getInt(groupPosition + positionOffset));
-                }
-            }
-
-            @Override
-            public void close() {
-                delegate.close();
-            }
-        };
+        return ValuesBytesRefAggregators.wrapAddInput(delegate, state, values);
     }
 $endif$
 
@@ -390,8 +261,8 @@ $if(long||double)$
         private final LongLongHash values;
 
 $elseif(BytesRef)$
-        private final LongLongHash values;
-        private BytesRefHash bytes;
+        final LongLongHash values;
+        BytesRefHash bytes;
 
 $elseif(int||float)$
         private final LongHash values;
